requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
cssbeautifier==1.15.1
cssmin==0.2.0
dj-database-url==2.2.0
Django==5.0.6
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-js-asset==2.2.0
djangorestframework==3.15.2
EditorConfig==0.12.4
gunicorn==22.0.0
htmlmin==0.1.12
idna==3.7
jsbeautifier==1.15.1
jsmin==3.0.1
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
PyPDF2==3.0.1
reportlab==4.2.2
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
